requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
Django==6.0.1
djangorestframework==3.16.1
faiss-cpu==1.13.2
filelock==3.20.3
fsspec==2026.1.0
google-ai-generativelanguage==0.10.0
google-api-core==2.29.0
google-auth==2.47.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.1
openai==2.15.0
packaging==25.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
proto-plus==1.27.0
protobuf==6.33.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.3.0
python-decouple==3.8
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.0
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.5
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.1
transformers==4.57.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
